Beefed up the ways in which trans_gloss_alignments are returned.
If it's only aligned at the gloss morph stage (as is the default), then create a word-based alignment and return THAT when asked. So directly compares with the other word-based one.

Also added the ability to look for a specific "created-by" type of alignment, so we can store multiple alignments in one file to compare.


Former-commit-id: f6d77d88ec48219bae19bbe916612e39676dbbae