Added unk_handling to the projection method from gloss to language line. This seems like a natural place to have it, since this function should by default have access to the gloss-to-language alignment.
Also did some bug fixes in how grams are supplied to the tagger and classifier -- some of the tokens in the XAML data have whitespace internal to the token, so those need to be removed if they are to be treated as a single token by the other tools.

Finally, added a number of the created_by= optional arguments to functions so that multiple tiers of annotation can coexist within a single document.


Former-commit-id: 9aedb20bd76aa0b6d40c33b47196c926a93596eb